+++ zcash_client_sqlite/src/wallet/init/migrations/tx_retrieval_queue.rs
@@ -5,8 +5,6 @@ use schemerz_rusqlite::RusqliteMigration;
 use std::collections::HashSet;
 use uuid::Uuid;
 
-#[cfg(feature = "transparent-inputs")]
-use zcash_primitives::transaction::Transaction;
 use zcash_primitives::transaction::builder::DEFAULT_TX_EXPIRY_DELTA;
 use zcash_protocol::consensus;
 
@@ -31,6 +29,7 @@ use {
     transparent::address::TransparentAddress,
     zcash_client_backend::data_api::DecryptedTransaction,
     zcash_keys::encoding::AddressCodec,
+    zcash_primitives::transaction::Transaction,
     zcash_protocol::{
         TxId,
         consensus::{BlockHeight, BranchId},