+++ zcash_client_backend/CHANGELOG.md
@@ -23,7 +23,6 @@ workspace.
   - `GetMempoolTxRequest` (previously named `Exclude`) has added field `pool_types`
 - `impl From<zcash_client_backend::proposal::ProposalError> for zcash_client_backend::data_api::wallet::input_selection::InputSelectorError<...>`
 - `zcash_client_backend::fees::MetaSource`
-- `zcash_client_backend::wallet::transparent` module, behind the `transparent-inputs` feature flag.
 - `zcash_client_backend::data_api::ll`
 
 ### Changed
@@ -73,7 +72,7 @@ workspace.
 
 ### Removed
 - `zcash_client_backend::data_api::testing::transparent::GapLimits` use
-  `zcash_client_backend::wallet::transparent::GapLimits` instead.
+  `zcash_keys::keys::transparent::GapLimits` instead.
 
 ## [0.21.0] - 2025-11-05