+++ zcash_client_sqlite/CHANGELOG.md
@@ -22,13 +22,16 @@ workspace.
 - `impl zcash_client_backend::data_api::ll::LowLevelWalletWrite for WalletDb`
 - `impl Hash for zcash_client_sqlite::{ReceivedNoteId, UtxoId, TxRef}`
 - `impl {PartialOrd, Ord} for zcash_client_sqlite::UtxoId`
+- `impl zcash_keys::keys::transparent::gap_limits::AddressStore for WalletDb`
+  (behind the `transparent-inputs` feature flag)
+- `zcash_client_sqlite::AccountRef` is now public.
 
 ### Changed
 - Migrated to `orchard 0.12`, `sapling-crypto 0.6`.
+- `zcash_client_sqlite::error::SqliteClientError` has added variant `GapAddresses`.
 
 ### Removed
-- `zcash_client_sqlite::GapLimits` use
-  `zcash_client_backend::wallet::transparent::GapLimits` instead.
+- `zcash_client_sqlite::GapLimits` use `zcash_keys::keys::transparent::GapLimits` instead.
 - `zcash_client_sqlite::UtxoId` contents are now private.
 
 ## [0.18.11, 0.19.2] - 2026-01-30